+++ lib/segment/tests/integration/scroll_filtering_test.rs
@@ -36,14 +36,16 @@ fn test_filtering_context_consistency() {
             &hw_counter,
             DeferredBehavior::Exclude,
         );
-        let read_by_stream_res = segment.filtered_read_by_id_stream(
-            Some(random_offset.into()),
-            Some(10),
-            &filter,
-            &is_stopped,
-            &hw_counter,
-            DeferredBehavior::Exclude,
-        );
+        let read_by_stream_res = segment.with_view(|view| {
+            view.filtered_read_by_id_stream(
+                Some(random_offset.into()),
+                Some(10),
+                &filter,
+                &is_stopped,
+                &hw_counter,
+                DeferredBehavior::Exclude,
+            )
+        });
 
         assert_eq!(read_by_index_res, read_by_stream_res, "filter: {filter:#?}");
     }
